initial commit of the 'getopts-long-rs' project
This is just a minimal crate intended to be published to crates.io
in order to hold the project namespace while the code gets
implemented.

And before you ask: Yes, I wanted to use the name 'getopt-long'
(without the ess), but the name was already taken.

Motivation:
-----------
While there are other Rust cli parsing libs, none of them seem right
for my purposes:

    - The 'getopt' library looks nice, but it's POSIX-only (no
      GNU-style long opts).

    - Using the 'getopts' lib (with an ess) on new projects is
      strongly discouraged (though I'm not sure that the lib itself
      is yet officially deprecated).

    - The 'pico-args' lib does not support bundling of short
      options. Would not be able to support, e.g., '-vvv'.

    - Most other libs I've looked at ('argh', 'clap', 'gumdrop',
      'structopt', 'getopt-long', others) want to play a role in
      crafting the help message displayed to the user. As an end
      user, I can usually tell when the help message has been
      generated rather than hand-crafted; the results are seldom
      pleasing. I have a strong distaste for such messages. They
      unnecessarily cast a shadow over the quality of the tool
      itself.